chore: refactor amplifyGeofenceControl (#73)
