+++ src/utils.ts
@@ -1,8 +1,5 @@
-import along from "@turf/along";
-import distance from "@turf/distance";
-import { lineString, point } from "@turf/helpers";
 import { Feature } from "geojson";
-import { AnySourceImpl, GeoJSONSource, LngLatBounds } from "maplibre-gl";
+import { AnySourceImpl, GeoJSONSource } from "maplibre-gl";
 import { Coordinates, NamedLocation, Geofence, Polygon } from "./types";
 
 export function isCoordinates(array: unknown): array is Coordinates {
